+++ HISTORY.md
@@ -1,4 +1,5 @@
-# Changelog
+Changelog
+=========
 
 - Adds support for Google Container Registry.
   [href]
@@ -12,7 +13,8 @@
 - Changes the cache path to `~/.cache/seantis/ostree` for non-root users.
   [href]
 
-# 0.0.0 (2018-08-24)
+0.0.0 (2018-08-24)
+------------------
 
 - Initial Release.
   [href]